Pad the background only once per segment
This removes some of the overhead of the padding process. Though it is
relatively quick, it may still take a few percent of the runtime.
Moreover, the previous implementation was rather sloppy with memory.

Memory usage has improved quite significantly for some input parameters.

In order to accomplish this, some rather cursed math with the slicing
was necessary.