Make the rotations divisor part of Config
This is in preparation of changes to the rotations mechanism such that
it has access to more fine-grained control over the number of structures
that are placed for a single rotation. That is _not_ implemented with
this commit. The present commit just serves to clean it up a bit
beforehand.