Replace voxelization routine with a Rust extension
This replaces the old voxelization routine written in Python with a Rust
function that computes a better voxelization, faster.

The new voxelization is better because cases where a bead's radius goes
beyond the borders of the voxel that contains a bead's center point, the
relevant neigboring voxels are also filled.

Before, the voxelization just filled the voxels that contained at least
one point. Cases where the point is just on the border of a voxel would
only fill one voxel, while perhaps four voxels together 'contain' the
bead if we take the radius into account.

Resolves #2 :)