Ensure at most the set number of rotations are placed
Before, it would be possible for more rotations to be tried than
configured. This is unpredictable and the behavior tends to be dependent
on how dense the system is: low-density systems would be more likely to
respect the provided number of rotations than high-density systems,
since more attempts would be made with a new rotation set for each one.