voxelization: Fix three beads at axactly the axis-zeroes
Because we move the points by the minimum value for x, y, z
respectively, at least three beads will lie exactly on those three
planes. That suggests that at least one and at most 9 voxels that
neighbor that bead in the negative domains in those spatial dimensions
should be filled in as well. But the voxel map only has a positive
domain.

Before, these voxels were not filled in.

Now, we adjust the `min` Vec3 by subtracting one bead radius, moving the
beads exactly out of the way of that problem described above.