pack: Re-use the background between sessions with identical compartments
If consecutive segment placement sessions have identical compartment
ids, the final state of the session background of the previous one is
identical to a session background that is computed from scratch for the
next one.

Previously, the session background would be recomputed for every
session, even if that was not necessary.
With this change, session backgrounds are re-used if possible.