pack: List nrots and more totals in the final summary
This commit presents the number of rotations that are stored for each
segment in the summary and adds more totals to the final line.

Before, only a total time was given in the final summary line.

Now, the total number of rotations, total target numbers, and total hits
are presented as well.