Rewrite Space::get_free_locations
Before this change, a great amount of runtime was spent in calculating
the three-dimensional coordinates from the linear coordinate provided by
the enumerate operation on the iterator (`make_idx`).
Around 45% of profiler samples were attributed to that section.

Since the three-dimensional coordinates and a linear coordinate can be
stored and incremented separately and trivially, this seemed a much more
efficient strategy for finding indices for open cells.

Turns out that this saves ~45% in time per segment for the 400 nm test
cell.