Rename the `number` field in Segment to `target`
The name `target` reflects the meaning more clearly and aligns the
wording with other sections, such as the summary.

Note that while the field for `state::Segment` is now called `target`,
the equivalent field in `config::Segment` is still called `number`. This
is to retain the equivalence to the structure of the input file.