Store open locations as a linear index
This reduces the memory footprint of storing the free locations
three-fold, since each location is stored as a `usize` now, instead of
the `[u64; 3]` like before.

The reduced memory footprint brings a performance increase by itself,
but the `linear_indices_where` function introduced in this commit
seems to present a significant improvement as well.