Only renew locations when necessary
Note that there is a critical flaw in the current design, as noted in
the source. The locations must be renewed whenever the set of
compartments changes between segment placements, but they are not as of
now. This means that it is assumed that consecutive segments have the
same set of compartments.

This is an intermediate commit that will be followed by a refactor that
will fix the issue.