Take a min and max value for checking rules
Additionally, the scalar value in `Position` rules is now described by a
`Scalar` type alias and it is now a `f32` instead of a `f64`.

Also, the `Rule::At` variant was removed. It may turn up later, but for
now it is convoluted to implement since it requires a centering system.