Fold render-placements into the bentopy source tree
Before, render-placements was developed in parallel with this repository
in a seperate source tree.

This commit folds the render-placements sources into this repository and
makes the modifications necessary to compile it here.

The Rust pyo3 extension for accessing render-placements from the python
entrypoint to bentopy has been removed.

In future commits, the `render` binary will see some internal changes,
where the eightyseven crate will be used as the gro reading and writing
back-end.