pack: Supply placement list output path as a command line argument
Before, the path would be generated from the title defined in the output
section of the input configuration, and the file would be placed in the
directory specified in that same output section.

This was convoluted and bad. The output directory was unnecessary, since
there is just one file. And to change the name of the placement list
that is generated, you had to change the input configuration.

Now, the path is specified through the command line arguments. No output
directory is created/used anymore.