Rework gro writing to be extremely memory-efficient
Before, the whole structure would be created in the `solvate` function,
by appending the accepted water beads onto the atoms of the global
structure. This global structure was then written out to the final gro
file.

But, this is not very memory efficient. To give a sense of this, at
sizes of around (400 nm)³, the memory footprint of the structure would
be around 22 GiB, with the `structures/water.gro` water box.

By only calculating the positions of the accepted water beads in chunks,
this issue is eliminated.

The present commit introduces this behavior.

Now, the most memory-intensive part of the process is the chunked
formatting of the atom lines for the output gro file. The chunk size can
be adjusted to fit a user's memory needs, if necessary.