check: Place collision bead at first bead, not between
This change makes inspecting the output of the check subcommand more
intuitive.

Before, when a collision would occur over periodic boundaries, the
collision bead would be placed in the middle of the box, making it hard
to understand between which structure beads the collision occurs.

Now, all structure beads that are involved with a collision are marked
by a bead at the same position.