Remove overly strict assertion
Before, this assert would trigger when the solvated box was smaller than
the solvation template in at least one dimension. Though this is an
unlikely scenario, it may occur. In fact it was triggered in a little
experiment with Bart last week.

As far as I can reason, it is a false positive, since the proper
translation over the boundary conditions is still applied.
Testing without issues has given me further confidence that the assert
led to false positives and can be removed without compromising
correctness.