pack: Rewrite the voxelize function to act robustly and conservatively
Before, the function was very susceptible to degenerate cases even with
uncommon yet reasonable settings, such as for high bead radius values or
very small voxel size resolutions. Since the old function only checked
the first shell of voxels, the voxelization produced was not
conservative for all settings.

This behaviour is fixed in this change. The function makes a much
stronger guarantee about conservative voxelization, now. This bolsters
bentopy's goal of producing clash-free initial conditions.

This change also gives a roughly 1.4× performance improvement for the
voxelize function. Note, however, that the total runtime contribution
of the voxelization is not very high (single digits percentage) to start
with, and that this improvement for the total bentopy-pack runtime is
therefore negligible.