pack: Move Mask bitwise operator logic to the mask module
Much more reasonable, right? Initially I had placed these
implementations in the rules module because they were only relevant
there and because the notion of a Rule was an experiment back then. But
as the project grew, we have started to use the bitwise operations
elsewhere and the notion of rules has become a mature part of bentopy.

Time to give them a more reasonable home.