pack: Restore ability to not rearrange components
This allows users to retain the manually specified ordering in the input
file.

In the past, this was possible, but since multiple rearrangement methods
were added, this option though supported by the internal logic could not
be set from the cli.

Now it can, through the addition of the `none` variant for the
arrangement method.