mask: Make full/free volume reporting robust
Before, cases existed where some label specifications, especially when
using `--autofill` would lead to errors during the np.unique unpacking.
In cases where all or no cells were selected, the counts could not be
unpacked into the two full and free categories, leading to an error.

This change expresses the desired counting of free and full cells in a
more clear and explicit manner that is not susceptible to cases where
all or no cells are selected.